requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
groq==0.8.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
lxml==5.2.2
oauth2client==4.1.3
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
pyparsing==3.1.2
pyTelegramBotAPI==4.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simplegmail==4.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
